requirements.txt
alembic
aniso8601
anyio
backports.cached-property
cffi
click
cryptography
Cython
ecdsa
exceptiongroup
fastapi
graphene
graphql-core
graphql-relay
greenlet
h11
httptools
idna
jose
lia-web
Mako
MarkupSafe
mysql-connector-python
packaging
passlib
pyasn1
pycparser
pydantic
Pygments
PyMySQL
python-dateutil
python-dotenv
python-gql
python-jose
python-multipart
PyYAML
rsa
six
sniffio
SQLAlchemy
starlette
starlette-graphene3
strawberry-graphql
tomli
typing_extensions
uvicorn
uvloop
watchfiles
websockets